Tidy up callbacks from ProgressTasks.
Shift more of the logic for handling pausing and resuming of
asynchronous tasks into the ProgressTask class, and update the
subclasses appropriately. The implementation is still ugly, but
it seems a bit better than before.

The problem is that we need to work with the latest copy of an
Application or Settings class in the code that is executed when
a ProgressTask completes. And these classes may have been
destroyed and recreated in the meantime (due to pause and
resume).